refractor: move DefaultValueConverter to internal project as its not yet supported
