+++ dev/io.openliberty.mcp/src/io/openliberty/mcp/annotations/ToolArg.java
@@ -29,8 +29,6 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 import java.util.Optional;
 
-import io.openliberty.mcp.tools.DefaultValueConverter;
-
 /**
  * Annotates a parameter of a {@link Tool} method.
  */
@@ -55,10 +53,8 @@ public @interface ToolArg {
     /**
      * The default value is used when an MCP client does not provide an argument value.
      * <p>
-     * {@link String}, primitive types and corresponding wrappers, and enums are converted automatically. For any other
-     * parameter type a custom {@link DefaultValueConverter} is needed.
-     *
-     * @see DefaultValueConverter
+     * {@link String}, primitive types and corresponding wrappers, and enums are converted automatically.
+     * Other types cannot have a default value.
      */
     String defaultValue() default "";
